feat(pins): add upvotes and downvotes to pins (#29)
Adds a new interaction to pins where users can upvote and downvote pins based on their reaction. When users click on a pin to view further details, they can click the up arrow to add an upvote, or the down arrow to downvote. They can click on the same icon again to remove/undo their choice. Other users would be able to view this interaction.